fix: code review suggestions
