+++ src/components/Signup/DefaultConfirmationContent/index.test.js
@@ -1,8 +1,9 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
+import user from '@testing-library/user-event';
 import { BrowserRouter as Router } from 'react-router-dom';
-import DopplerIntlProvider from '../../i18n/DopplerIntlProvider.double-with-ids-as-values';
-import { DefaultConfirmationContent } from './DefaultConfirmationContent';
+import DopplerIntlProvider from '../../../i18n/DopplerIntlProvider.double-with-ids-as-values';
+import { DefaultConfirmationContent } from '.';
 
 describe('DefaultConfirmationContent', () => {
   it('should render warning message when resentTimes is more than 0', async () => {
@@ -29,14 +30,14 @@ describe('DefaultConfirmationContent', () => {
     );
 
     // Assert
-    expect(screen.queryByTestId('warning-message')).toBeInTheDocument();
+    expect(screen.getByText('signup.no_more_resend_MD_link')).toBeInTheDocument();
   });
 
-  it('should render the user email', async () => {
+  it('should call the incrementAndResend function if resend email button is clicked', async () => {
     // Arrange
-    const incrementAndResend = () => {};
+    const incrementAndResend = jest.fn();
     const registeredUser = 'harcodedUser@mail.com';
-    const resentTimes = 1;
+    const resentTimes = 0;
     const mailtoSupport = 'mailtoSupport@mail.com';
     const Captcha = () => null;
 
@@ -56,6 +57,8 @@ describe('DefaultConfirmationContent', () => {
     );
 
     // Assert
-    expect(screen.getByText(registeredUser)).toBeInTheDocument();
+    const button = screen.getByRole('button');
+    await user.click(button);
+    expect(incrementAndResend).toHaveBeenCalled();
   });
 });
